pyproject.toml
[tool.poetry]
name = "delta_utils"
version = "0.4.0"
description = ""
authors = ["Niels Lemmens <1936122+bulv1ne@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/husqvarnagroup/delta_utils/"
documentation = "https://delta-utils.readthedocs.io/en/latest/"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "*"
reverse_geocoder = "*"
iso3166 = "*"
delta-spark = "*"

[tool.poetry.dev-dependencies]
mkdocs-material = "^8.2.9"
mkdocstrings-python = "^0.8.2"
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
mkdocs = "*"
moto = {extras = ["s3"], version = "*"}
boto3-stubs = {extras = ["s3"], version = "*"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."mkdocs.plugins"]
RootFiles = "mkdocs_plugins.files:RootFiles"
